Programm fertig (hoffentlich)

+++ ContactManager_ZBW/View_Cyril/Form1.cs
@@ -1,22 +1,14 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-using ContactManager_ZBW.Milos.Controller;
+﻿using ContactManager_ZBW.Milos.Controller;
 using ContactManager_ZBW.Model_Renato;
+using System;
+using System.Windows.Forms;
 
 
 namespace ContactManager
 {
     public partial class mainview : Form
     {
-        private Controller Controller { get ; set; }
+        private Controller Controller { get; set; }
 
         public mainview()
         {
@@ -30,9 +22,9 @@ namespace ContactManager
             throw new NotImplementedException();
             // soll hier schon geladen werden?
             LoadList();
-            
+
         }
-        
+
         // Function CmdNew_Click
         // description: creates a new Person and adds it to the list
         private void CmdNew_Click(object sender, EventArgs e)
@@ -42,6 +34,7 @@ namespace ContactManager
                 Person temporaryPerson = CreateCustomerOrEmployee();
                 FillAllFields(temporaryPerson);
                 int index = Controller.CreateNewPerson(temporaryPerson);
+                LoadList();
                 LslContactList.SelectedIndex = index;  // select and show newly created person
                 if (index != -1)
                 {
@@ -51,14 +44,15 @@ namespace ContactManager
                 {
                     MessageBox.Show("Person bereits vorhanden.");
                 }
-            }             
+            }
         }
 
         // Function CmdUpdate_Click
         // description: after a Person is selected, its specs can be altered with this function
         private void CmdUpdate_Click(object sender, EventArgs e)
         {
-            if (LslContactList.SelectedIndices.Count > 0) {
+            if (LslContactList.SelectedIndices.Count > 0)
+            {
                 if (CheckNecessaryFields()) // only when necessary Fields are entered
                 {
                     {
@@ -66,6 +60,8 @@ namespace ContactManager
                         Person temporaryPerson = CreateCustomerOrEmployee();
                         FillAllFields(temporaryPerson);
                         Controller.UpdatePerson(index, temporaryPerson);
+                        LoadList();
+                        LslContactList.SelectedIndex = index;   // select and show updated person
                     }
                 }
                 else
@@ -100,17 +96,24 @@ namespace ContactManager
         private void CmdDelete_Click(object sender, EventArgs e)
         {
 
-                if (LslContactList.SelectedIndices.Count > 0)
+            if (LslContactList.SelectedIndices.Count > 0)
+            {
+                DialogResult result;
+                result = MessageBox.Show("Möchten Sie diese Person wirklich löschen?", "Löschvorgang", MessageBoxButtons.YesNo);
+                if (result == DialogResult.Yes)
                 {
-                    throw new NotImplementedException; // Noch Abfragefenster programmieren!
                     int index = LslContactList.SelectedIndex;
                     Controller.DeletePerson(index);
                     ClearView();
+                    LoadList();
+                    MessageBox.Show("Person gelöscht.", "Löschvorgang");
                 }
-                else
-                {
-                    MessageBox.Show("Keine Person zum Löschen ausgewählt.");
-                }            
+
+            }
+            else
+            {
+                MessageBox.Show("Keine Person zum Löschen ausgewählt.");
+            }
         }
 
         // Function CmdSave_Click
@@ -134,7 +137,7 @@ namespace ContactManager
         // description: when a person is selected in the list, its data will be shown in the view
         public void LslContactList_SelectedIndexChanged(object sender, EventArgs e)
         {
-            if (LslContactList.SelectedIndices.Count !=0 && LslContactList.SelectedIndices.Count < 2)
+            if (LslContactList.SelectedIndices.Count != 0 && LslContactList.SelectedIndices.Count < 2)
             {
                 int selectedIndex = LslContactList.SelectedIndex;
                 Person selectedPerson = Controller.GetPerson(selectedIndex);
@@ -151,25 +154,25 @@ namespace ContactManager
 
             if (TabControl.SelectedTab.Name == "TbCustomer")
             {
-                temporaryPerson = new Customer;
+                temporaryPerson = new Customer();
             }
             else // if (TabControl.SelectedTab.Name == "TbEmployee")
             {
                 if (RdTraineeYes.Checked)
                 {
-                    temporaryPerson = new Trainee;
+                    temporaryPerson = new Trainee();
                 }
                 else
                 {
-                    temporaryPerson = new Employee;
+                    temporaryPerson = new Employee();
                 }
             }
             return temporaryPerson;
         }
 
-    // Function CheckNecessarFields
-    // description: checks if all necessary fields are entered before actions
-    private bool CheckNecessaryFields()
+        // Function CheckNecessarFields
+        // description: checks if all necessary fields are entered before actions
+        private bool CheckNecessaryFields()
         {
             // Check if all necessary base fields are entered:
             if (TxtFirstName.Text != "" && TxtLastName.Text != "" && TxtEmail.Text != "")
@@ -185,7 +188,7 @@ namespace ContactManager
 
         // Function FillAllFields
         // description: fills all the fields of the view into the Person object
-        private void FillAllFields (Person person)
+        private void FillAllFields(Person person)
         {
             // fill all base fields of the person:
             person.Salutation = RdSalutationsFemale.Checked ? "Frau" : "Herr";
@@ -218,7 +221,11 @@ namespace ContactManager
             {
                 case "TbCustomer":
                     ((Customer)person).CompanyName = TxtCompanyName.Text;
-                    ((Customer)person).CustomerType = TxtCustomerType.Text;
+                    ((Customer)person).CustomerType = Convert.ToChar(TxtCustomerType.Text);
+                    if (!Controller.CheckCustomerType(((Customer)person).CustomerType))
+                    {
+                        MessageBox.Show("Bitte nur Werte von A-F in Kundentyp eingeben.");
+                    }
                     ((Customer)person).CompanyContact = TxtCompanyContact.Text;
                     break;
                 case "TbEmployee":
@@ -233,21 +240,21 @@ namespace ContactManager
                     }
                     catch (FormatException)
                     {
-                        MessageBox.Show("Beschäftigungsgrad und Kaderstufe dürfen nur Nummern sein.");                        
+                        MessageBox.Show("Beschäftigungsgrad und Kaderstufe dürfen nur Nummern sein.");
                     }
                     // also get trainee data if neccessary:
                     if (RdTraineeYes.Checked)
                     {
                         try
                         {
-                            ((Trainee)person).ActualTraineeYear = Convert.ToInt16(TxtActualTraineeYear.Text);
-                            ((Trainee)person).TraineeYears = Convert.ToInt16(TxtTraineeYears.Text);
+                            ((Trainee)person).ActualTraineeYear = Convert.ToInt32(NumActualTraineeYear.Value);
+                            ((Trainee)person).TraineeYears = Convert.ToInt32(NumTraineeYears.Value);
                         }
                         catch (FormatException)
                         {
-                            MessageBox.Show("Lehrjahre und aktuelles Lehrjahr dürfen nur Zahlen sein.");                                                        
+                            MessageBox.Show("Lehrjahre und aktuelles Lehrjahr dürfen nur Zahlen sein.");
                         }
-                    }                                     
+                    }
                     break;
                 default:
                     break;
@@ -261,7 +268,7 @@ namespace ContactManager
         private void ShowAllFields(Person person)
         {
             // show all base fields of the person:
-            RdSalutationsFemale.Checked = person.Salutation=="Frau"? true :false;
+            RdSalutationsFemale.Checked = person.Salutation == "Frau" ? true : false;
             TxtFirstName.Text = person.FirstName;
             TxtLastName.Text = person.LastName;
             DtpDateOfBirth.Value = person.DateOfBirth;
@@ -301,14 +308,14 @@ namespace ContactManager
                 TxtRole.Text = ((Employee)person).Role;
                 DtpStartDate.Value = ((Employee)person).StartDate;
                 DtpEndDate.Value = ((Employee)person).EndDate;
-                TxtEmployment.Text = Convert.ToString( ((Employee)person).Employment );
-                TxtCadreLevel.Text = Convert.ToString( ((Employee)person).CadreLevel );
+                TxtEmployment.Text = Convert.ToString(((Employee)person).Employment);
+                TxtCadreLevel.Text = Convert.ToString(((Employee)person).CadreLevel);
 
                 //additionally show all trainee fields in the view:
                 if (person is Trainee)
                 {
-                    TxtTraineeYears.Text = ((Trainee)person).TraineeYears;
-                    TxtActualTraineeYear.Text = ((Trainee)person).ActualTraineeYear;
+                    NumTraineeYears.Value = ((Trainee)person).TraineeYears;
+                    NumActualTraineeYear.Value = ((Trainee)person).ActualTraineeYear;
                 }
             }
 
@@ -341,23 +348,24 @@ namespace ContactManager
         {
             if (LslContactList.SelectedIndices.Count > 0)
             {
-               
+
                 int index = LslContactList.SelectedIndex;
-                Customer customer = Controller.GetPerson(index);
-                
+                Person customer = (Customer)(Controller.GetPerson(index));
+
 
 
                 string newLogEntry = TxtLogNew.Text;
                 if (newLogEntry != "")
                 {
                     // Noch über Controller machen, nicht direkt im Model.
-                    throw new NotImplementedException() // Noch index und so machen;
+                    throw new NotImplementedException(); // Noch index und so machen;
                     Customer.AddLogEntry(index, newLogEntry);
+
+                    customer.AddLogEntry(DateTime.Now, newLogEntry);
+                    Controller.UpdatePerson(index, customer);                 // Update new data to controllers customer-list
+                    TxtLogNew.Clear();                      // and clear entry field
+                    LslContactList.SelectedIndex = index;   // finally update view and show entry in list
                 }
-                customer.AddLogEntry(DateTime.Now, newLogEntry);
-                Controller.UpdatePerson(customer);                 // Update new data to controllers customer-list
-                TxtLogNew.Clear();                      // and clear entry field
-                LslContactList.SelectedIndex = index;   // finally update view and show entry in list
                 else
                 {
                     MessageBox.Show("Keine leeren Logeinträge erlaubt.");
@@ -375,6 +383,11 @@ namespace ContactManager
         }
 
         private void TxtLogNew_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void TxtTraineeYears_TextChanged(object sender, EventArgs e)
         {
 
         }
